fix: dynamicValue in directive list
