+++ packages/jsx-compiler/src/modules/directive.js
@@ -1,5 +1,6 @@
 const t = require('@babel/types');
 const traverse = require('../utils/traverseNodePath');
+const dynamicValueHelper = require('../utils/dynanmicValueHelper');
 const genExpression = require('../codegen/genExpression');
 
 const directiveIf = 'x-if';
@@ -91,38 +92,10 @@ function transformDirectiveCondition(ast) {
   return dynamicValue;
 }
 
-function transformDirectiveList(ast) {
+function transformDirectiveListAttr(ast) {
   const dynamicValue = {};
-  traverse(ast, {
-    JSXElement: {
-      exit(path) {
-        const { node } = path;
-        const { attributes } = node.openingElement;
-        if (node.__jsxlist) {
-          const { args, iterValue } = node.__jsxlist;
-          dynamicValue[iterValue.name] = iterValue;
-          attributes.push(
-            t.jsxAttribute(
-              t.jsxIdentifier('a:for'),
-              t.stringLiteral(`{{${genExpression(iterValue)}}}`)
-            )
-          );
-          args.forEach((arg, index) => {
-            attributes.push(
-              t.jsxAttribute(
-                t.jsxIdentifier(['a:for-item', 'a:for-index'][index]),
-                t.stringLiteral(arg.name)
-              )
-            );
 
-            // Mark skip ids.
-            const skipIds = node.skipIds = node.skipIds || new Map();
-            skipIds.set(arg.name, true);
-          });
-          node.__jsxlist = null;
-        }
-      }
-    },
+  traverse(ast, {
     JSXAttribute(path) {
       const { node } = path;
       if (t.isJSXIdentifier(node.name, { name: 'x-for' })) {
@@ -165,12 +138,50 @@ function transformDirectiveList(ast) {
   return dynamicValue;
 }
 
+function transformDirectiveList(ast) {
+  const dynamicValue = {};
+  traverse(ast, {
+    JSXElement: {
+      enter(path) {
+        const { node } = path;
+        const { attributes } = node.openingElement;
+        if (node.__jsxlist) {
+          const { args, iterValue } = node.__jsxlist;
+          const skipIds = node.skipIds = node.skipIds || new Map();
+          skipIds.set(iterValue.name, true);
+
+          dynamicValueHelper.setValueInDirectiveList(path, iterValue, dynamicValue);
+          attributes.push(
+            t.jsxAttribute(
+              t.jsxIdentifier('a:for'),
+              t.stringLiteral(`{{${genExpression(iterValue)}}}`)
+            )
+          );
+          args.forEach((arg, index) => {
+            attributes.push(
+              t.jsxAttribute(
+                t.jsxIdentifier(['a:for-item', 'a:for-index'][index]),
+                t.stringLiteral(arg.name)
+              )
+            );
+            // Mark skip ids.
+            skipIds.set(arg.name, true);
+          });
+          node.__jsxlist = null;
+        }
+      }
+    }
+  });
+  return dynamicValue;
+}
+
 module.exports = {
   parse(parsed, code, options) {
     if (parsed.renderFunctionPath) {
       Object.assign(
         parsed.dynamicValue = parsed.dynamicValue || {},
         transformDirectiveCondition(parsed.templateAST),
+        transformDirectiveListAttr(parsed.templateAST),
         transformDirectiveList(parsed.templateAST),
       );
     }
